separated to /home endpoint to split page into components
